Adding new base commands system
Creating a way to integrate basic attacks using cards. This can be changed later, but seems like the easy way to make attack actions, as they would take up the same 'slot' as other cards and can be resolved like any other card.

BaseCommands handles the display of the new cards, and currently, only a goth attack card exists to test it. Will try and make this work tonight